Move controllers for resources to resources package rather than config-entries package
