+++ control-plane/resources/controllers/consul_resource_controller_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package controllersv2
+package controllers
 
 import (
 	"context"
