+++ control-plane/resources/controllers/grpc_route_controller.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package controllersv2
+package controllers
 
 import (
 	"context"
